[add fix] tmp_views.pyの記述を各機能ごとのviewsファイルに 分割して格納、また"id"に使用されていた”_”を削除
